Add Testcontainers-based end-to-end integration test + fix latent bugs it flushed
New tests/MFilesExporter.IntegrationTests project:
- Testcontainers.MsSql boots SQL Server 2022 once per collection.
- SqlServerFixture provisions the tracking DB (applies committed DDL
  scripts 10, 15, 20, 30, 35, 40, 70, 71; skips 50-security since sa is
  in use; skips 60-maintenance pending a separate archive-sproc fix)
  plus a synthetic vault schema (DOCUMENTFILEVERSION / DATAFILEVERSION /
  DATAFILEVERSION_BYTES) matching the canonical query.
- VaultSeeder inserts N documents with varying payload sizes and
  deterministic SHA-256 checksums.
- ExporterTestHost composes the same DI graph as Program.cs but points
  at the container and a temp output directory.
- ExportSmallCorpusTests.ExportsAllDocuments_WithMatchingChecksums runs
  the full pipeline against 100 seeded documents and asserts:
    * every seeded document lands on disk exactly once,
    * every file's SHA-256 matches its expected value,
    * tracking-DB counters (recorded/succeeded/failed/skipped/bytes)
      reconcile with the corpus.

Latent bugs the test surfaced and fixed:
- SqlStreamingEngine + MFilesSqlDocumentEnumerator called GetInt64 on
  the ID_VERSIONPART column, which is INT in the M-Files vault schema.
  This would have thrown InvalidCastException on every real production
  read. Changed to GetInt32 with widening to long.

Known follow-up (not fixed here):
- 60-maintenance.sql's archive sproc uses SELECT j.* into
  archive.ExportJobs which has one fewer column (no RowVersion). Skipped
  in the fixture; TODO to fix separately since it does not affect the
  pipeline path.

Test runtime: ~1 s warm, ~30 s cold (SQL container boot).

+++ src/MFilesExporter.Persistence/MFiles/MFilesSqlDocumentEnumerator.cs
@@ -96,7 +96,8 @@ internal sealed class MFilesSqlDocumentEnumerator : IDocumentEnumerator
         while (await reader.ReadAsync(cancellationToken).ConfigureAwait(false))
         {
             var docPart = reader.GetInt64(0);
-            var verPart = reader.GetInt64(1);
+            // ID_VERSIONPART is INT in the vault schema.
+            var verPart = (long)reader.GetInt32(1);
             var title = await reader.IsDBNullAsync(2, cancellationToken).ConfigureAwait(false)
                 ? string.Empty : reader.GetString(2);
             var ext = await reader.IsDBNullAsync(3, cancellationToken).ConfigureAwait(false)
